requirements.txt
pyyaml==6.0.2
numpy==1.24.4
networkx==3.1
scipy==1.10.1
pandas==2.0.3
opencv-python==4.12.0.88
tqdm==4.67.1
openai-clip==1.0.1
scikit-learn==1.3.2
graphviz==0.20.3
wget==3.2
open-clip-torch==2.32.0
torchvision==0.19.0
torch==2.4.0
